Remove circular dependency between ethers-event.ts and internal-types.ts
