+++ src/internal/ethers-event.ts
@@ -1,3 +1,5 @@
+import { EventType, Filter, Listener } from '@ethersproject/abstract-provider';
+
 import {
   AlchemyEventFilter,
   AlchemyEventType,
@@ -10,9 +12,83 @@ import {
 import {
   ALCHEMY_EVENT_TYPES,
   ALCHEMY_MINED_TRANSACTIONS_EVENT_TYPE,
-  ALCHEMY_PENDING_TRANSACTIONS_EVENT_TYPE,
-  Event
+  ALCHEMY_PENDING_TRANSACTIONS_EVENT_TYPE
 } from './internal-types';
+
+/**
+ * DO NOT MODIFY.
+ *
+ * Event class copied directly over from ethers.js's `BaseProvider` class.
+ *
+ * This class is used to represent events and their corresponding listeners. The
+ * SDK needs to extend this class in order to support Alchemy's custom
+ * Subscription API types. The original class is not exported by ethers. Minimal
+ * changes have been made in order to get TS to compile.
+ */
+class Event {
+  readonly listener: Listener;
+  readonly once: boolean;
+  readonly tag: string;
+
+  _lastBlockNumber: number;
+  _inflight: boolean;
+
+  constructor(tag: string, listener: Listener, once: boolean) {
+    this.listener = listener;
+    this.tag = tag;
+    this.once = once;
+    this._lastBlockNumber = -2;
+    this._inflight = false;
+  }
+
+  get event(): EventType {
+    switch (this.type) {
+      case 'tx':
+        return this.hash!;
+      case 'filter':
+        return this.filter!;
+      default:
+        return this.tag;
+    }
+  }
+
+  get type(): string {
+    return this.tag.split(':')[0];
+  }
+
+  get hash(): string {
+    const comps = this.tag.split(':');
+    if (comps[0] !== 'tx') {
+      throw new Error('Not a transaction event');
+    }
+    return comps[1];
+  }
+
+  get filter(): Filter {
+    const comps = this.tag.split(':');
+    if (comps[0] !== 'filter') {
+      throw new Error('Not a transaction event');
+    }
+    const address = comps[1];
+
+    const topics = deserializeTopics(comps[2]);
+    const filter: Filter = {};
+
+    if (topics.length > 0) {
+      filter.topics = topics;
+    }
+    if (address && address !== '*') {
+      filter.address = address;
+    }
+
+    return filter;
+  }
+
+  pollable(): boolean {
+    const PollableEvents = ['block', 'network', 'pending', 'poll'];
+    return this.tag.indexOf(':') >= 0 || PollableEvents.indexOf(this.tag) >= 0;
+  }
+}
 
 /**
  * Wrapper class around the ethers `Event` class in order to add support for
